Move cloud code tests
