+++ spec/cloud/cloudCodeAbsoluteFile.js
@@ -1,0 +1,3 @@
+Parse.Cloud.define('cloudCodeInFile', (req, res) => {
+  res.success('It is possible to define cloud code in a file.');
+});
