+++ src/cloud-code/Parse.Cloud.js
@@ -1,4 +1,4 @@
-import { Parse } from 'parse/node';
+import { Parse }     from 'parse/node';
 import * as triggers from '../triggers';
 
 function validateClassNameForTriggers(className) {
@@ -40,11 +40,15 @@ ParseCloud.afterDelete = function(parseClass, handler) {
   var className = getClassName(parseClass);
   triggers.addTrigger(triggers.Types.afterDelete, className, handler, Parse.applicationId);
 };
-  
+
 ParseCloud._removeHook = function(category, name, type, applicationId) {
   applicationId = applicationId || Parse.applicationId;
   triggers._unregister(applicationId, category, name, type);
 };
+
+ParseCloud._removeAllHooks = () => {
+  triggers._unregisterAll();
+}
 
 ParseCloud.httpRequest = require("./httpRequest");
