+++ src/triggers.js
@@ -1,6 +1,6 @@
 // triggers.js
-import Parse from 'parse/node';
-import AppCache  from './cache';
+import Parse    from 'parse/node';
+import AppCache from './cache';
 
 export const Types = {
   beforeSave: 'beforeSave',
@@ -49,13 +49,17 @@ export function removeTrigger(type, className, applicationId) {
    delete _triggerStore[applicationId].Triggers[type][className]
 }
 
-export function _unregister(a,b,c,d) {
-  if (d) {
-    removeTrigger(c,d,a);
-    delete _triggerStore[a][b][c][d];
+export function _unregister(appId,category,className,type) {
+  if (type) {
+    removeTrigger(className,type,appId);
+    delete _triggerStore[appId][category][className][type];
   } else {
-    delete _triggerStore[a][b][c];
+    delete _triggerStore[appId][category][className];
   }
+}
+
+export function _unregisterAll() {
+  Object.keys(_triggerStore).forEach(appId => delete _triggerStore[appId]);
 }
 
 export function getTrigger(className, triggerType, applicationId) {
